Adding new field and logic to messaging
Added in COMPDTTM to the view and needed that to appear in the app. Also
based the logic behind messaging on this new field so that if there is
no compdttm the message in the comments field will appear, if there is a
date than the pre-defined messaging will come up.